Day 11 (Part 2): Fighting with the cache...
I got hinted that a better cache should solve the problem... but all the cache key that tried were inefficient.
Even I tried to convert to return the all the paths... Good that I could not make that work because I would have run out of memory...
The problem that took me ages to understand was that I was using the visited set that I used to avoid loops to determine if the path is valid... but using that means the cache cannot be used...

Trying to explore other avenues y started returning the 4 cases (valid, with dac only, with fft only and with none) and with a little bit of careful joining it worked like a charm.

P.S. Never use ints to hold results. You should have learned the lesson by now.