Day 10 (Part 2): One of the horrible problems so far...
And that's a wrap for one of the most horrid problems I had to solve. I went for the system solver which needed for me to take a crash course on lineal systems solving using gauss reductions. I haven't done this stuff since the 90ties

First I hoped I could calculate this using combinations with repetition... until the numbers were crazy. So I wasted a day in figuring out an algorithm.

Then I checked reddit for hints and went for the linear equation system.

And Java does not have any magic like the python z3 so I have to rely on a stack overflow response for the row echelon calculations, then it took me ages to figure out how to determine the free variables (I had to switch from BigDecimal to double because the stack overflow code did not take into account the crazy precision of BigDecimal...)
I have to use a reference implementation for compare my free variables calculation and when I finally manage it to match, solving the system was a horrible fight int and doubles...
But done...
I did not enjoy this AoC that much. Not many crazy problems and just 4 o 5 interesting and this one was not much enjoyable for me. I re-learned a lot of Algebra in order to solve this